system: moved codeInsight and highlighting abstractions into system
