+++ src/com/haskforce/system/integrations/highlighting/HaskellAnnotationHolder.java
@@ -1,6 +1,5 @@
 package com.haskforce.system.integrations.highlighting;
 
-import com.intellij.codeInsight.daemon.impl.AnnotationHolderImpl;
 import com.intellij.lang.annotation.Annotation;
 import com.intellij.lang.annotation.AnnotationHolder;
 import com.intellij.lang.annotation.HighlightSeverity;
